Implemented text generation based on candidate, added level counter, did level transitioning with Henry
Made it so that the outcome texts is generated based on candidate rather
than hardcoded.

Added a level counter in EvandriaUpdate to identify which level the
player is on.

Did Level Transitioning and persistent data with Henry. Henry will
implement level transitioning and I did persistent data and we will
integrate soon